Clean up userInfo util usage and props
